fix: use uint64 for disk size bytes

+++ internal/store/sqlite/store_test.go
@@ -25,7 +25,7 @@ func TestStoreUpsertDiskAndAppendEvent(t *testing.T) {
 		Model:      "WD Red",
 		Serial:     "WD-123456",
 		Transport:  "sata",
-		SizeBytes:  4000787030016,
+		SizeBytes:  uint64(4000787030016),
 		Rotational: true,
 	}
